Better error handling
Added error handling on failed execution with a custom exception that
bears most of the info regarding the failed query.

Also added a new FluentUtils function that takes a query and a bunch
of parameters and outputs them roughly the same way they'd be queried,
that's very useful for debugging.